Former-commit-id: 15e0609f4ea416abe699490dc8f81426d46ed2e9

+++ backend/mindderStart/src/main/java/com/ssafy/mindder/scraps/controller/ScrapsController.java
@@ -10,6 +10,7 @@ import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
@@ -18,6 +19,7 @@ import com.ssafy.mindder.common.ErrorCode;
 import com.ssafy.mindder.common.SuccessCode;
 import com.ssafy.mindder.common.dto.ApiResponse;
 import com.ssafy.mindder.feeds.controller.FeedsController;
+import com.ssafy.mindder.feeds.model.FeedListDto;
 import com.ssafy.mindder.feeds.model.FeedsParameterDto;
 import com.ssafy.mindder.scraps.model.service.ScrapsService;
 import com.ssafy.mindder.util.JwtService;
@@ -29,14 +31,14 @@ import io.swagger.annotations.ApiParam;
 @RestController
 @RequestMapping("/scraps")
 public class ScrapsController {
-	
+
 	@Autowired
 	private ScrapsService scrapsService;
 	@Autowired
 	private JwtService jwtService;
 
 	private static final Logger logger = LoggerFactory.getLogger(FeedsController.class);
-	
+
 	@ApiOperation(value = "스크랩 등록", notes = "스크랩을 등록한다.")
 	@PostMapping("/{feedIdx}")
 	public ApiResponse<?> scrapAdd(@RequestParam("access_token") String accessToken, @RequestParam int feedIdx) {
@@ -44,7 +46,9 @@ public class ScrapsController {
 		logger.debug("scrapAdd - 호출 : " + feedIdx);
 		try {
 			int userIdx = jwtService.getUserIdx(accessToken);
-
+			if (scrapsService.findScrap(userIdx, feedIdx) != null) {
+				return ApiResponse.error(ErrorCode.VALIDATION_SCRAP_EXCEPTION);
+			}
 			scrapsService.addScrap(userIdx, feedIdx);
 			return ApiResponse.success(SuccessCode.CREATE_SCRAP);
 		} catch (Exception e) {
@@ -53,7 +57,7 @@ public class ScrapsController {
 			return ApiResponse.error(ErrorCode.INTERNAL_SERVER_EXCEPTION);
 		}
 	}
-	
+
 	@ApiOperation(value = "스크랩 삭제", notes = "피드 번호에 해당하는 스크랩을 삭제한다.")
 	@DeleteMapping("/{feedIdx}")
 	public ApiResponse<?> scrapRemove(@RequestParam("access_token") String accessToken,
@@ -62,7 +66,9 @@ public class ScrapsController {
 		logger.debug("scrapRemove - 호출 : " + feedIdx);
 		try {
 			int userIdx = jwtService.getUserIdx(accessToken);
-			
+			if (scrapsService.findScrap(userIdx, feedIdx) == null) {
+				return ApiResponse.error(ErrorCode.NOT_FOUND_SCRAP_EXCEPTION);
+			}
 			scrapsService.removeScrap(userIdx, feedIdx);
 			return ApiResponse.success(SuccessCode.DELETE_SCRAP);
 		} catch (Exception e) {
@@ -71,16 +77,15 @@ public class ScrapsController {
 			return ApiResponse.error(ErrorCode.INTERNAL_SERVER_EXCEPTION);
 		}
 	}
-	
+
 	@ApiOperation(value = "스크랩 목록 조회", notes = "유저 번호에 해당하는 피드의 목록을 반환한다.", response = FeedsParameterDto.class)
 	@GetMapping("/my")
-	public ApiResponse<?> myScrapList(@RequestParam("access_token") String accessToken) {
+	public ApiResponse<?> myScrapList(@RequestHeader("access_token") String accessToken) {
 
 		logger.debug("myScrapList - 호출 : ");
 		try {
 			int userIdx = jwtService.getUserIdx(accessToken);
-
-			List<FeedsParameterDto> scrapList = scrapsService.findMyScraps(userIdx);
+			List<FeedListDto> scrapList = scrapsService.findMyScraps(userIdx);
 			return ApiResponse.success(SuccessCode.READ_MY_SCRAP_LIST, scrapList);
 		} catch (Exception e) {
 			e.printStackTrace();
